changed chainId to BigInt in TxRequest and TxReceipt, and adjusted CLI tests to reflect changes

+++ packages/cli/src/__tests__/project/sample.ts
@@ -205,7 +205,7 @@ type Ethereum_TxRequest @imported(
   gasPrice: BigInt
   data: String
   value: BigInt
-  chainId: UInt32
+  chainId: BigInt
   type: UInt32
 }
 
@@ -265,7 +265,7 @@ type Ethereum_Network @imported(
   nativeType: "Network"
 ) {
   name: String!
-  chainId: Int!
+  chainId: BigInt!
   ensAddress: String
 }
 
@@ -424,7 +424,7 @@ type Ethereum_TxResponse @imported(
   gasPrice: BigInt
   data: String!
   value: BigInt!
-  chainId: UInt32!
+  chainId: BigInt!
   blockNumber: BigInt
   blockHash: String
   timestamp: UInt32
@@ -498,7 +498,7 @@ type Ethereum_TxRequest @imported(
   gasPrice: BigInt
   data: String
   value: BigInt
-  chainId: UInt32
+  chainId: BigInt
   type: UInt32
 }
 
@@ -792,7 +792,7 @@ type Ethereum_TxRequest @imported(
   gasPrice: BigInt
   data: String
   value: BigInt
-  chainId: UInt32
+  chainId: BigInt
   type: UInt32
 }
 
@@ -852,7 +852,7 @@ type Ethereum_Network @imported(
   nativeType: "Network"
 ) {
   name: String!
-  chainId: Int!
+  chainId: BigInt!
   ensAddress: String
 }
 
@@ -869,7 +869,7 @@ type Ethereum_TxResponse @imported(
   gasPrice: BigInt
   data: String!
   value: BigInt!
-  chainId: UInt32!
+  chainId: BigInt!
   blockNumber: BigInt
   blockHash: String
   timestamp: UInt32
